Temporarily fix container width bug on About page
If there's not a lot of content on a page, the layout width expands,
affecting not only the content but the rest of the page, inclding
the headers. The proper fix should be somewhere in the CSS. This
fix will just allow us to sidestep the issue while we address more
pressing matters.