[#19] Conditionally include primary/secondary cols
We can transition templates to single-column layouts without
changing their parents or the parents of related templates in
the same group by adding an empty single_column block
    {% block single_column %}{% endblock %}
to the template that establishes the single-column layout. When
that block is present, the page.html template will not render the
primary and secondary columns. If it's not present, then the
primary and secondary columns will be included as usual. This
prevents changes to one template from breaking other templates.