Show login/register page to anonymous beta users
Together with ckanext-oauth2, this prevents unauthorized users
from viewing all pages except for the static ones. They're
presented with an explanation of the situation and links to log
in or register. If oauth2 is enabled and configured to use the
NextGEOSS UM, the user will be redirected to the UM site for
authentication.