Show login/register page to anonymous beta users
Together with ckanext-oauth2, this prevents anonymous users from
viewing all pages. They will be presented with an explanation of
the situation and links to log in or register. If oauth2 is
enabled and configured to use the NextGEOSS UM, clicking those
links will redirect users to the UM site for authentication.