Handle string extras conversions in one place (#64)
* Handle string extras conversions in one place

Handle all the string extras conversions in the IPackageController
hook. The extras are converted from a string to a list of dicts
by package_show and package_search and before being indexed, so
the package dicts that templates, etc. receive have normal extras
and there's no need to add helpers or write special functions to
process the string extras in each extension, template, etc.

* Clarify purpose of hooks